Improve list-marker positioning
* Fix horizontal position to be outside the border box of the parent
instead of outside the content box
* Align bullet markers with the font baseline so that the vertical position
adapts better to different fonts and sizes
* Handle vertical alignment of `outside` list markers to adapt their
position for larger line heights. They now also affect the height of the
corresponding line box
* Apply word and letter spacing for text markers

The following issue is still unresolved:

* Inside positioned text markers do not properly take the space they
need. This is due to the fact that the exact marker text is only
resolved in the renderer. To fix this, the marker text would need to be
resolved in the reflow process, similar to counters (actually, most of
the logic could be shared).

Fixes #478
Fixes #708
Fixes #1801
Fixes #1902
Fixes #2451